All-GitHub-Repos/pyproject.toml
[project]
name = "all-github-repos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.1.0",
    "pygithub>=2.8.1",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
]
